Slight improvement to search function on index.php
I have managed to
1. Change the code in js/index.js to Jquery, and added a function where the search results automatically show as the user types in his words.
2. Change the design of the table that outputs the search results.

Point 1 is a success. But point 2 has created problems;
a. I am unable to change the width of the table, and the table design is vastly different from what I saw on Bootstrap's website. I believe it might be a problem with using echo to print the html code from displayVocabBook.php.
b. The collapsible in displayVocabBook.php that hides/shows the meaning of the stored words broke. It is able to show the contents when hidden, but is unable to hide the contents when shown. I removed the modal I added, thinking it caused the break, but it was not the cause. I believe it might be the table design from Bootstrap.

Current proposal to solve problem (b);
Perhaps return the search results in an array, then use index.js to print them into a table.
Or find a way to fix the html code when it is echoed from displayVocabBook.php back to index.js.

Proposal to improve table design:
I guess I could change the table to Accordion? It looks like a nice way to present the word meaning (and other data too once I add those features).

This is all I can think of at the moment.