Used Bootstrap template to improve looks of website
I am using a Bootstrap template with the hopes of improving the looks of my website.
To make the website work, I am making some changes to the js and php scripts, but the function of the website will remain the same - for people to store words they learn.

Currently, I am trying to make the login form submit to index.js where it gets forward via ajax to login.php. However, instead of submittined through index.js as a POST, I always see the parameters in the URL (meaning a GET is used). I am not able to solve this problem yet.